src/requirements.txt
azure-functions
aiohttp
requests
python-dotenv